--- tests/conftest.py
@@ -1,2 +1,0 @@
-TEST_IMP_KEY = "imp_apikey"
-TEST_IMP_SECRET = "ekKoeW8RyKuT0zgaZsUtXXTLQ4AhPFW3ZGseDA6bkA5lamv9OqDMnxyeB9wqOsuO9W3Mx9YSJ4dTqJ3f"
